Skeletons rarely drop skeleton keys
From UnNetHack. The motivation for this is purely a little pun based
mechanic.

The frequency is even lower than in UnNetHack; I don't think it needs to
be particularly common and if skeletons are ever added earlier than
Orcus, they shouldn't be treated as a likely source of a key.