Fix: ghosts could be hit by any weapon
This seems to be a regression caused by a vanilla merge at some point,
because in a game I was playing I was able to very slowly wear down and
kill a ghost with a uncursed rusty +1 long sword. There was a
shade-specific check negating the damage from non-shade_glare() objects;
correct that to noncorporeal().