Identify cancellation if it blows up a bag of holding
From UnNetHack. This could be viewed as a simple addition to the
unambiguous auto-wand-IDing that xNetHack does, since no other wand
blows up bags of holding.

I also added two features: it also identifies the bag of holding (since
no other type of bag is prone to explode), and gives a respectable
amount of experience points. This latter feature is because it will be
hilarious when someone blows up their bag of holding and gains a level
as a consolation prize.