Add level sounds for branch-staircase levels
One common complaint in NetHack is that it's hard to find the entry
point of a branch across a range of possible levels where it can
generate. Particularly Vlad's Tower, since the stairway is randomly
hidden in the maze and can be quite tedious to find, but also happens
when the Mines staircase generates in a well-hidden room.

So I've added some. Like most level sounds, they have both normal and
hallucinatory counterparts. Additionally, while the Mines-stairs sound
will continue to recur, the Sokoban and Vlad ones get turned off when
one Sokoban level is finished and Vlad is killed. I could have made it
so that they stopped appearing once you had visited the level on the
other side of the staircase, but since you can just pop in and then
leave, it seemed better to tie it to some measure of completion of that
level.