Fix: properly handle "The X of Y" artifacts in string parsing
When I changed the Caveman artifact to Big Stick, I did not realize I
was violating some assumptions - that all quest artifacts, and only
quest artifacts, would have "The X of Y" names.

Since I don't really want to keep that assumption intact, I added a
function that tells whether an artifact's name begins with "The", and
called it in the appropriate places.