Replace the Orb of Detection with Itlachiayaque
This is part of the Archeologist quest redesign and was in my backlog,
but was actually not something I realized was needed until I was
rewriting the quest text and had to decide between either doing this now
or shoehorning the Orb of Detection in where it wouldn't really fit
anymore.

So: Itlachiayaque. It comes from a patch by L, via UnNetHack, and is
broadly similar to those implementations. It is a shield of reflection
(its material is gold, because the depictions I can find of it refer to
it being "golden", plus much of the planned loot in the
Archeologist quest is items made of gold and this fits in well with
that). It confers fire resistance and warning when carried, and
Archeologists only can wear it without taking a spellcasting penalty.

Like in the original patch, it can be invoked for a stinking cloud, but
I also gave it a second invoke power that the player can select from:
since in its mythology it was used as a scrying device to look at
distant places, here you can gaze into it and get the effects of a
crystal ball. This also parallels the Orb of Detection nicely, but
notably the amount of times you can do it with Itlachiayaque is
unlimited and does not require any charging.

(Also note that the stinking cloud doesn't work quite the same as in
UnNetHack; here I tried to avoid duplicating the code that makes a
stinking cloud and put it in a common function, so it mostly works like
a scroll of the same beatitude, with the same range limit. But the
effect where a non-Archeologist automatically gets the cloud centered on
themselves is the same.)

In UnNetHack, it retains half spell damage, but doesn't have that here:
I dropped half spell damage because in general I think that an artifact
should provide only one of half spell, half physical, reflection, or
magic resistance. The Orb provided two but was largely useless otherwise
so it got a bit of a pass. And yes, I'm aware that other quest artifacts
currently provide two.

As part of this commit, I turned resists_fire into a proper function; it
is impossible to set two monster extrinsics in update_mon_intrinsics(),
which Itlachiayaque now does. The proper way to do this appears to be
via functions like resists_magm that can check for artifact effects on
top of the regular effects.