Fix: wrong messages for region dissipation
Noticed that I got a "The gas cloud enveloping you dissipates" message
when standing just outside a cloud, adjacent to it on the top and right.
This is because I was still in the rectangle of the region, and that's
all that was being checked. It needs to also check inside_region for
correctness.

Ditto for the case where you're not in the cloud but get a message for
seeing it dissipate.