Replace the Minion of Huhetotl with Schliemann
Another key component of the Archeologist quest redesign. A big part of
the quest redesigns is to get away from the quest nemesis always being a
generic big bad type who ran off with the quest artifact twirling their
mustache and snickering. The Minion was already one of the blander
nemeses (I mean, it doesn't even have a name, and is basically a Nalzok
clone) so this is an opportunity to inject some better flavor.

Schliemann is based on Heinrich Schliemann, a nineteenth-century amateur
archeologist controversial for his destructive methods. I'm not totally
happy with the name just being "Schliemann", but using his first name is
wordy, "Dr. Schliemann" is inaccurate, and "Herr Schliemann" doesn't
sound right either.

Combat-wise, he is similar to other human quest nemeses, having no
particular magical power (apart from wearing Itlachiayaque) and having
decent but middling attacks. To spice things up a bit I gave him a +4
pick-axe to wield and, to make it feel like a more unique boss fight, a
bunch of oil potions that he can now throw at the player (it's the
closest thing NetHack has to explosives, and the Arc quest redesign
involves his team blasting their way down to the quest artifact).
As of this commit, he still inexplicably warps to the stairs and has the
power to heal himself in d8 increments when not near the hero. (Ideally
he would still pursue Itlachiayaque, but not warp or heal.) I have also
given him the same behavior as centaurs have, where he won't try to
engage the player directly in melee, unless the player is carrying
Itlachiyaque in which case he'll attack directly.

This commit does not change any quest dialogue or the goal level; that
will come later in the Archeologist quest redesign.