Fix: use correct behavior for des.line() with one coordinate
One would expect that a line from e.g. (10,10) to (10,10) would contain
only that one point, but I discovered that lspo_line doesn't know what
to do with that input and slices a big diagonal line across the whole
map in both directions from that point. This corrects the behavior to
stop after selecting the one point.

This problem exists in vanilla NetHack, but is latent because des.line
is never used.