Allow covetous monsters to follow normal AI
This is a bit of an experiment, because I am not happy with Schliemann
following normal covetous behavior. Apart from holding Itlachiyaque, he
shouldn't have any magic available, and the planned design for the goal
level doesn't make any sense for him to start warping to the upstairs
(he wouldn't have been in that part of the level yet).

This adds a macro that attempts to disable covetous behavior in cases
where it pertains to warping and special movement behavior, without
affecting the other ramifications of the flag such as not allowing them
to be pacified. This included a special case in the nemesis dialogue
because apparently that code is reliant on the nemesis not being next to
you and then warping to you...

Downside: Covetous nonwarpers will not appear to go after the coveted
object at all. E.g. if it's sitting on the floor, or held by another
monster, they won't go there or attack the monster; they'll just keep
attacking the player. For quest nemeses, where neither of these
situations is likely to happen much, this is acceptable right now. I
think it would take a larger AI rewrite to allow nemeses to chase their
target without the ability to instantly arrive there.