Monsters can invoke Itlachiyaque for a gas cloud
This is kind of unprecedented, as I'm not aware of any other variant
allowing monsters to use the invoke powers of artifacts. This extends
only to Itlachiayaque (no other artifact invokes are implemented for
monsters), and the purpose is to add yet another dimension to the Arc
boss fight.

I placed it in the miscellaneous AI code because offensive item use
requires the target to be lined up, but summoning a gas cloud should
only requires line of sight. (I didn't explicitly reimplement the radius
here; miscellaneous item use already requires the hero be within a
certain radius.) The monster must also be able to see the hero.

Additionally, Schliemann can invoke it more often than the usual invoke
timeout, so that he might use it more than as just an opening salvo in
the fight with him. The handwave explanation for this is that he's
forcing the artifact to be used before it's ready, without concern for
the consequences.