Accept "waiting" on special level des.monster() specifications
This is a pretty good alternative to specifying monsters as sleeping,
because I wanted some monsters that don't start immediately wandering
towards the hero once the level is generated, but I didn't want to
specify them as asleep either because then the hero could walk right
past without waking them up. The WAITFORU code is perfect for this, but
it otherwise needed to be specified in the permonst, which isn't any
good - I only wanted these particular monsters to behave like this. By
letting it be set in the level parser, I get the best of both worlds.