Overhaul the Archeologist quest levels and flavor text
This is the bulk of the Arc quest overhaul I have alluded to in earlier
commits. None of the levels were particularly compelling to start with,
and the quest story, while certainly not the *worst* NetHack has to
offer, was not compelling either, being just another variation on the
same old "the nemesis has stolen our artifact, and you need to go get it
back".

I've been reimagining the stories for several quests to get away from
this formulaicity, and this is the first one I've taken a crack at. The
new story is based on such archaeological adventure stories as Raiders
of the Lost Ark and National Treasure. Now, the College in the home
level isn't under attack, and never had Itlachiayaque in the first place
-- they merely have discovered where it is. But the nemesis has caught
wind of this and is going there to snatch it first, and *he* doesn't
care about blowing up some old tomb if there's treasure in it. You're
tasked with getting to it first. (But of course, you're not quick
enough, and they have found Itlachiayaque by the time you get there;
otherwise it wouldn't be all that interesting. Plus, you need to get the
Bell from them anyway.)

The lower filler levels use an all-new level generator that I
implemented completely in Lua, as a bit of a proof of concept that the
level generation algorithm doesn't have to be implemented in C. In those
levels, and the second half of the locate and the first half of the
goal, I was going for a distinctly tomb-like feel with a lot of hidden
areas and secrets to discover, and where it's not trivial for the hero
to just walk through and proceed down the next staircase.

Monsters are changed up (though as of this commit, monster *generation*
is not -- so the levels deep in the ancient unopened tomb will still be
spawning a lot of snakes over time, ugh). Apart from some wild animals
in the jungle areas, most of the threat on this quest is from mummies.

Loot is changed up as well - it won't be *quite* as loot-heavy of a
quest since there aren't as many snakes generated at level creation
time. There aren't scattered random objects anymore - most of the loot
in the tomb is items made of gold (and other exotic materials), though I
added a number of chests which will still provide random contents.