Allow webs to be placed without a giant spider on them
This is something I've found very annoying. Up until now, placing a web
always meant placing a giant spider; this was hardcoded in the trap
creation. This meant that webs could not be used in some places such as
early levels where giant spiders are too strong, or in levels where
flavor-wise there really shouldn't be any spiders. It's the latter case,
for the Archeologist quest redesign, that is the impetus here: there was
no way to have random traps in the tomb without them also coming with
giant spiders that don't belong in such a place.

There were several options for how to do it, and this is the most
complicated but also the most precise, allowing the level designer to
control whether a spider generates on a web, for individual traps and
not with a level-wide rule.

Since there was a FIXME in the Ozymandias themed room about this very
feature not existing yet, I changed that too.