Fix: running on grass/ice sent you in weird directions
qt reported this. It's because grass was being treated as a type of
corridor rather than a ROOM space, and running has rules for following
similar terrain. In reality it's not a corridor and should behave like
normal room for running purposes.

Also do this for ice, because this same thing happens in the valkyrie
quest ice fields, which is a vanilla bug.