Fix: don't show messages for AD_PITS attack in mhitm cases
create_pit_under() didn't correctly account for the fact that the
function might be called on two monsters nowhere near the hero, so it
would print these messages seemingly from nowhere. Add canseemon()
checks to account for that (the messages describe specifics of what the
monsters are doing which seems more specific than what canspotmon() will
get you.)

Based on a fix to this same problem in EvilHack but I implemented it
differently.