Allow rubbing silver items to cure withering
The gameplay motivation for this is, as I've said in an earlier commit,
the desire to have some cure for withering that is not as rare as prayer
nor as infinite and ubiquitous as unicorn horns. The non-gameplay
inspiration is the short story Shadows for Silence in the Forests of
Hell, in which silver has similar effects on supernatural withering as
it does here - except in the game, it only cures the Wither status, and
doesn't restore any of the HP lost, because the amount of HP lost to
withering isn't tracked and it would be a clumsy system if it tried to.

The main thing I was trying to avoid with this system was one in which
it's optimal to have a big stack of lightweight silver items such as
darts, and rubbing them one by one cancels a greater amount of withering
timeout than rubbing the whole stack at once. Eventually I had to do
away with a bit of code that gave a little extra healing if the item was
erodeproofed, because I couldn't come up with a way around this problem.

I was also indecisive on whether silver was the right material choice
here; silver is already a highly sought after material and generally
piling even more uses onto it is not good - but there aren't a ton of
other candidates. Platinum is really the only one that would make sense,
and it's a lot rarer (in order to be used much, the hero would need to
collect most pieces of platinum gear they came across) and less
flavorful, with silver having the traditional healing properties.