Block giving monster your iron ball while punished
An iron ball that is chained to a punished hero is not supposed to end
up in a monster's inventory.  Monsters are already prevented from
picking the ball or chain up off the ground, but a player who picked it
up could use #loot to give it to a pet.  This could cause problems with
various pieces of ball and chain code, which expect it to be impossible
for the ball and chain to end up in a monster's inventory (or several
other places), and could cause various panics and crashes.  Prevent this
from happening by blocking attempts to give 'uball' or 'uchain' to a
monster via #loot.