Improve feedback for giving partial stack to pet
Feedback given on failure to give a monster part of a stack via #loot,
for instance because handling the object would stone them or because
they didn't have room in their inventory to take it, would not
respect the quantity of items specified by the user.  Change messages in
this kind of case to use "any of", e.g. "The giant refuses to take
any of your cockatrice corpses."

Also provide additional feedback for cases where the recipient monster
can only take part of the stack, so that it's clearer why your little
dog only successfully gets one item after you tried to give it more than
one.