Permit giving an item to a saddled pony
Because a monster with no hands was blocked from being given any more
items via #loot if it was carrying something already, and a saddle (or
other worn item) is placed in the monster's inventory, a saddled
monster like a pony could not be given any items besides the saddle.
Prevent worn items in the monster's inventory from counting against the
inventory limit of one item.