Fix: pet unable to "return" item to shop
Picking up an itemin a shop, using #loot to give it to your pet, then
allowing your pet to drop the item, would cause the item to remain on
your bill despite it having been returned to the shop inventory. If a
pet drops an unpaid item in a shop, ensure it is removed from the bill.