Fix: giving pet wielded/quivered item via #loot
Giving an equipped item (wielded, quivered, etc) to a pet didn't clear
its "currently equipped" flags or reset the appropriate uwep-type
pointer, which meant that it was still marked as wielded even once it
was in your pet's inventory. This could also trigger an impossible if
the player then used #loot to retrieve the item, which would reset its
owornmask but not change uwep (etc), and then attempted to wield another
item, swap weapons, etc.