Handle welded equipment in #loot more consistently
When using #loot to exchange items with a monster, it was impossible to
give or take various cursed equipped items, even if the item was of a
type that did not normally weld to your hand.  Use the welded and
mwelded functions to handle weapon welding in a way more like the rest
of the game.