Fix: swimming pool themed room min difficulty was too low
Complaint was that the pool generated on dungeon level 1 in a place that
required the player to walk around the edge to advance through the
level, and then got killed by the jellyfish in the pool, which... fine,
that's a reasonable complaint. Push it to never appear before dungeon
level 5.