Fix: gold detection didn't print messages or allow hero to view map
This is because in the initial broken implementation of passive gold
detection (a LONG time ago, and it still isn't fixed), for some reason I
can't figure out, reading a gold detection scroll called gold_detect
with passive = TRUE. This is just straight wrong; the gold detection
scroll is specifically active detection. Flipping it back to FALSE seems
to make it work properly again.