Fix: peacefuls would never take any items given via #loot
Due to an overlooked condition in can_carry that prevents peacefuls from
taking any items at all, which is intended so that they don't just
vacuum up items on the floor before the hero can get there. Add an
exception for items in the hero's inventory (the only time a peaceful
should be getting can_carry called for a hero-inventory item is when the
player in wizard mode is giving it to them) that permits them to take
it.