Fix: ghosts attempting to scare already paralyzed hero
Could happen either from being mobbed by ghosts, or from having a ghost
show up under the influence of a paralysis potion. Either way, it
doesn't make a ton of sense for the ghost to just sit there vanishing
and then popping back up while the player is paralyzed, and it was
certainly wrong to have "You are frightened to death and unable to move"
show up.

The fix is twofold: first, ghosts will not uninvisible themselves if the
player is paralyzed. This prevents the "appears out of nowhere" message
spam. Second, scary_ghost will no longer print anything or induce
paralysis if already paralyzed; I initially didn't have that in the
commit because I thought the ghost's active attack was the only way it
could get called on the monsters' turn, but then realized it could
happen if the paralyzed hero is teleported by a quantum mechanic into an
abandoned temple.