Fix: mimics imitating altars always showed as unaligned
Where in reality they're supposed to be capable of appearing as any
altar. I took this opportunity to finally get rid of an unmerged code
path between vanilla and xNetHack regarding altar colors: the vanilla
code was never reached. Now I've deleted the xNetHack code so it does
get reached, and merged it into the vanilla code (which already checked
correctly for mimics).

I decided to revert the flipping of chaotic and unaligned altar colors
back to what vanilla uses, but not lawful, which is still yellow: white
and gray _ often render as the same thing, and I consider having them
render as distinct colors more important than matching the altar colors
to their respective unicorns.