Make metal boxes impervious to destruction by #force
This is partly in prep for an upcoming feature, but also common sense -
metal tends to deform rather than break and is quite durable; no amount
of bashing with a hard object is likely to "totally destroy" a metal box
to the point of unusability where it's not even worth treating as an
object anymore. Similarly, the contents are probably not under any
threat.