Add mummy trapped chests
This went through a few iterations of how exactly to best do it. The
initial idea was that there would be a floor trap, with a valuable item
placed on it (imagine the opening scene from Raiders of the Lost Ark).
Stepping onto the trap, as with any trap, would trigger a swarm of
mummies to wake up and surround the player. However, this was easily
defeated: what if you kicked the item off? Used a grappling hook? Got a
gelatinous cube to pick it up? In fact, monsters should trigger this,
and if they did that would make it kind of boring for the hero, right?

So instead of a floor trap, it is a chest trap, since that neatly limits
the scope to things the player can do. It spawns exclusively on gold
chests in the Arc quest; I guess technically a metallivore could still
eat the chest but that destroys its contents so it's not abusable.

Internally this extends the spe field on chests to represent a chest
that if trapped should give mummies. These chests are also impossible to
untrap by normal means (MKoT will still work, but how many archeologists
have that on the quest?)

The prizes in the locate level chests are specific; mummy trap chests
that appear on other levels of the quest just have random contents.

Also, I was using fire traps to stand in for mummy traps in the locate
level, but now that it was time to remove them I had grown to like them.
After all, random gouts of flame are a standard feature in most temples
of this type. So there are a few of them placed randomly inside.