Change Arc quest monster generation probability
The main reason for this is to prevent the quest from becoming
absolutely broken in terms of loot when M is blessed-genocided. Because
the quest was pretty mummy-heavy, levels generated after genociding M
would have all the fixed mummies in the quest replaced with quest-random
monsters... which would be mostly snakes... including potentially large
groups of garter snakes... which all generate items on their square if
they generate at level creation time. Testing revealed that around 60
additional items per level could be created in this manner if you were
lucky.

By changing the probabilities (to favor mummies as the main thing that
gets generated, with snakes still present as a minority) this effect is
mostly ameliorated. Generating quest levels after genociding mummies
still yields some snakes, but not a ton, with the bulk of replacement
monsters being fully random.