Fix: properly support artifacts warning of monster classes
When Dragonbane and Trollsbane were buffed to include warning of dragons
and trolls respectively, it was done with a kludge that overloaded
g.context.warntype.obj with the monster index, and a lot of the code
around warning (enlightenment feedback, farlook) wasn't updated
accordingly. Additionally, I recently got a bug report that when
wielding Trollsbane, the player was actually warned against all the M2
flags that make up the bit representation of S_TROLL, including humans,
which was definitely unintended.

This was a known issue at the time it was implemented, but required a
savebreak to do properly. In this commit I have done the proper,
savebreaking fix, which adds a new field to warntype for "object that
warns of monster species" and updates the other bits of code that
involve warning. There are some assumptions and limitations of this
system described in comments, but it should work for Dragonbane and
Trollsbane.

This IS a save-breaking change.