Give the Oracle a potion of hallucination sometimes
Adds a slight bit of flavor to where all those cryptic messages might be
coming from and also ties into the history of the real-life Delphic
Oracle.

Initially the plan was to place it under her, but that was weird because
when you went out of sight, the potion would be shown at that spot, and
I feel like it might make players want to kill her more. So she gets it
in her inventory.