Fix: wished-for items with unspecified erosion shouldn't be eroded
Another bug induced by allowing items to generate eroded. The wish
parser assumed that calling mksobj with init=TRUE would never produce
erosion, so it didn't previously need to clear eroded flags on the item.
Intended behavior is the same: the only erosion that should happen on a
wished item is if the wisher explicitly asked for it.