Fix: charge to #loot pet outside shop
Taking an item from a pet while standing in a shop doorway would charge
you for the item as though the pet were standing in the shop.

The current behavior of looting a pet inside a shop -- that giving an
item to a pet immediately relinquishes it to the shop inventory, and
the player must pay to get it back -- is still a bit strange, in my
opinion, but this commit only addresses situations where the pet being
looted is not actually inside the shop.

Also properly mark 'no charge' items while they are in a monster's
inventory (this may happen if hero drops item in shop, declines to sell
it, pet picks it up, then hero loots pet).