Record completion of the quest in xlogfile
Chris_ANG suggested "completed the overhauled Arc quest" as a unique
xNetHack trophy for Junethack, and bhaak said the easiest way of
tracking it in Junethack is via the achieveX field. I noticed that the
vanilla xlogfile doesn't actually express quest completion everywhere,
but I think they may end up adding it as the 31st achievement so I
didn't try to add it as a formal achievement. Quest completion will not
be recorded in the achieve bitmask, just as a string in achieveX.

This has the additional effect of perhaps incentivizing players to play
archeologist up through the quest during the tournament, which I would
like to encourage. I'm not sure I'll keep this around permanently, but
for now it records this.