Fix: serverseed varied between first and subsequent game sessions
Due to there being a large random number added in newgame(), which
wasn't saved at all, when the game was reloaded it would then use
just sysopt.serverseed without this number. This affects not just
readable Hawaiian shirts but also things like which trap is on a door,
so I needed to fix it in some way given Junethack is running.

Temporary fix is to remove the large random number so that serverseed is
just in effect all the time. A proper fix would require a savebreak.