Fix: crash when monster read scroll of earth
... due to m_useup being called twice on the same object. This is
another casualty of the vanilla merge, in which my vanilla pull request
code (pulled locally into xNetHack because it fixed an immediate crash)
differed from PatR's modified version when adopting it into vanilla, and
the two by some quirk of git did *not* conflict with each other. Since
the divergence is minor, delete the xNetHack code in favor of vanilla's.