Fix: permahallu and permadeaf inverted in conductX in xlogfile
I can only assume this is a logical error, as all the other similar
conducts use a negation. But roleplay options shouldn't be negated.

Also add unfairscareless to conductX; I feel like I had some reason for
not putting it in there before but the commit message doesn't indicate
it either way.