Fix: thiefstones not updating container weight
Using a thiefstone to deposit an item inside a container did not
increase the total weight of the container.  As a result, it was
perfectly possible to take advantage of this to produce a sack stuffed
chock-full of magical items and treasure, which nonetheless had the
weight of an empty bag.