Fix: encumbrance message from removing armor
...appeared a turn late.  Because of the weight reduction of worn armor,
removing a piece of armor increases the hero's current total carried
weight.  If this pushes the weight into next encumbrance class
(e.g. from unburdened to burdened), the statusline reflected the
change immediately but the "your movements are slowed slightly by your
load" pline was not printed until the start of the next turn.