Fix: thiefstone vs very large objects
Boxes, boulders, and large statues are typically impossible to add to a
container: if a player tries to place one into a box or bag, they are
told that the item can't fit.

However, when using a thiefstone to deposit items at a stash location,
even items that were normally 'too large' were placed into a container
without any fuss.  A chest could be made to contain another chest or a
statue of a dragon, provided the items were golden and therefore
eligible for use with a thiefstone.

Apply the same rules as are used in in_container(pickup.c) for items
that are 'too large' to fit in a container; for anything which falls
into this category, teleport it as usual but place it onto the floor
at the keyed location rather than attempting to deposit it into a
container.