Fix: "The leather jacket welds itself" when putting on plastic jacket
This is because of a collision with the code that is intended to force
the default material on certain items that had their materials stripped
from the actual object, such as the jacket. This probably also happened
on similar armor items.

Fix by not showing any materials on items for any code path that goes
through distant_name (which this weld message does via minimal_xname and
simpleonames). I left an exception for statues; their material should be
clear even if you see them from a distance.