Clarify comments around 'balk' math for pet attacking logic
The math is correct, but has confused some people reading the code into
thinking that it allows the pet to attack a monster 3 levels higher than
it, which is not the case.

Also, the comment was apparently wrong in declaring that pets will never
attack anything if their health is under 25%; this only applies in the
case that they are considering attacking a peaceful. Thus, the 20%
threshold is still relevant and should be noted in the comment.