Fix: dmonsfree errors when pit fiend's created pit killed monster
When this happened (reported for tame pit fiend but I think it happened
if the player was polyselfed into one as well), the mhitm_ad_pits
function did not report the defender had died and had already been
cleaned up (thitm calls xkilled or monkilled as appropriate), so the
large-scale monster movement logic treated the monster as still capable
of doing things that turn before it realized it was dead and tried to
clean it up again. This is fixed by checking for the monster being dead
before the actual pit attack damage is applied, and returning
MM_DEF_DIED appropriately.

Also fix a related minor bug in which a flying monster killed by a pit
could print a message saying it "flies back up" after its death.