Revert drawbridge glyph/color change
This was an old change to make them more distinctive, turning them to +
because they're a sort of door. However, it's ended up just one of those
weird variant interface changes that no one likes; no one was actually
confusing raised drawbridges with anything else. So put it back the way
it was, to brown #.

Sinks are still blue \ because gray # is very confusable with other
glyphs.